pyproject.toml
[project]
name = "smartsuite-crud-chatbot"
version = "0.1.0"
description = "An AI Agent that performs CRUD in SmartSuite"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.37.0",
    "openai>=1.95.1",
    "openai-agents>=0.2.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
